Rename polls file to singular following docs
